src/pyproject.toml
[tool.poetry]
name = "shopping-items"
version = "0.1.0"
description = ""
authors = ["Bhushan Pathak <bhushan.s.pathak@ril.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.70.1"}
SQLAlchemy = "^1.4.29"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.12b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
